structure and draw floor ok
