+++ Makefile
@@ -25,7 +25,7 @@ CFILES_BONUS := main_bonus.c init_bonus.c utils_bonus.c free_memory_bonus.c \
                 validation_bonus/validation_bonus.c validation_bonus/check_rgb_bonus.c validation_bonus/data_processing_bonus.c validation_bonus/map_bonus.c validation_bonus/map_utils_bonus.c validation_bonus/check_wall_bonus.c \
                 graphic_bonus/hooks_bonus.c graphic_bonus/movements_bonus.c graphic_bonus/images_bonus.c graphic_bonus/draw_squares_bonus.c graphic_bonus/wall_bonus.c \
                 graphic_bonus/algorithm_bonus.c graphic_bonus/setup_bonus.c graphic_bonus/draw_loop_bonus.c \
-				graphic_bonus/minimap_bonus.c
+				graphic_bonus/minimap/minimap_bonus.c graphic_bonus/minimap/minimap_images_bonus.c
 
 # PATH_FILES
 SRCS        := $(addprefix $(SRC_PATH)/, $(CFILES))
