+++ includes_bonus/cub_bonus.h
@@ -19,7 +19,7 @@
 # define HEIGHT 600
 # define BLOCK 8
 # define MINI_WIDTH 200
-# define MINI_HEIGHT 100
+# define MINI_HEIGHT 200
 # define ROTATE_LEFT -0.5
 # define ROTATE_RIGHT 0.5
